# acceleration/api/openapi.yaml
openapi: 3.0.3
info:
  title: Model Router
  version: 0.2.0
  description: >
    Routes speech-to-text and text-to-speech traffic across providers and reports
    what it cost. Every path is scoped by modality, so the same provider serving
    two modalities is reported on separately. The caller is identified by a
    trusted X-Customer-Id header; real authentication is not part of this version.
servers:
  - url: http://localhost:8080
security:
  - CustomerId: []
paths:
...
  /v1/{modality}/providers:
    get:
      operationId: listProviders
      summary: List the providers configured for a modality and their live health
      parameters:
        - $ref: "#/components/parameters/Modality"
      responses:
        "200":
          description: The configured providers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Provider"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/{modality}/routes/{target}:
    get:
      operationId: resolveTarget
      summary: Resolve a provider name or capability shortcut to a ranked candidate list
      parameters:
        - $ref: "#/components/parameters/Modality"
        - name: target
          in: path
          required: true
          description: A "provider/model" name or a capability shortcut such as en-low-latency.
          schema:
            type: string
          example: en-low-latency
        - name: language
          in: query
          required: false
          description: Language hints that candidates must cover. Repeat for several.
          schema:
            type: array
            items:
              type: string
          example: ["en"]
      responses:
        "200":
          description: Candidates in preference order, best first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Candidate"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/{modality}/stats:
    get:
      operationId: getStats
      summary: Aggregated usage for the calling customer
      parameters:
        - $ref: "#/components/parameters/Modality"
        - name: granularity
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/Granularity"
        - name: from
          in: query
          required: true
          description: Start of the window, inclusive.
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          required: true
          description: End of the window, exclusive.
          schema:
            type: string
            format: date-time
        - name: tag
          in: query
          required: false
          description: >
            Only count requests carrying every one of these cost labels, each written
            "key:value". Repeat for several. Filtering reads the request rows rather
            than the rollups, since a rollup bucket no longer knows which labels its
            requests carried.
          schema:
            type: array
            items:
              type: string
          example: ["project:moderation", "environment:dev"]
      responses:
        "200":
          description: One row per bucket, provider and model
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StatsBucket"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/{modality}/stats/tags:
    get:
      operationId: getTagStats
      summary: Aggregated usage broken down by the values of one cost label
      description: >
        What drives the spend. Requests are labelled with whatever keys the customer
        chooses, so asking for key=project returns one row per project per bucket.
      parameters:
        - $ref: "#/components/parameters/Modality"
        - name: key
          in: query
          required: true
          description: The cost label to group by.
          schema:
            type: string
          example: project
        - name: granularity
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/Granularity"
        - name: from
          in: query
          required: true
          description: Start of the window, inclusive.
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          required: true
          description: End of the window, exclusive.
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: One row per bucket and label value
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TagStatsBucket"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/turns/stats:
    get:
      operationId: getTurnStats
      summary: Conversational latency for the calling customer
      description: >
        One row per bucket and agent. A request row measures one provider call; a turn
        measures what the caller felt, from finishing a sentence to hearing the answer
        start, with the transcription, model and voice legs kept apart so a slow
        conversation can be attributed.
      parameters:
        - name: agent_id
          in: query
          required: false
          description: Narrow to one agent. Omit for every agent the customer runs.
          schema:
            type: string
        - name: granularity
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/Granularity"
        - name: from
          in: query
          required: true
          description: Start of the window, inclusive.
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          required: true
          description: End of the window, exclusive.
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: One row per bucket and agent
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TurnStatsBucket"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /v1/agents/sessions/{id}/events:
    get:
      operationId: watchSession
      summary: Watch the conversation and answer the model's tool calls
      description: >
        A WebSocket, which OpenAPI cannot describe past the upgrade. Frames are JSON
        objects carrying a `type` and the fields of that event.

        The server sends what the conversation did: `joined`, `heard`, `responding`,
        `response_delta`, `responded`, `spoke`, `turn`, `decision`, `delegated`,
        `task_settled`, `task_cancelled`, `tool_call`, `tool_ran`, `transferred`,
        `pressed`, `looked_up`, `backchannel`, `interrupted`, `overlap_decided`,
        `conversation_compacted`, `error` and `left`.

        A `decision` frame is one judgement the conversation made, carrying the same
        fields as a CallEvent. Together they are why the call went the way it did, and
        they are also written down, so a finished call replays them from
        `/v1/agents/calls/{id}/events`.

        Two frames are only sent when asked for, because they are far more frequent than
        the rest and most consumers want neither. `interim=true` adds `hearing`, which is
        a transcript revision as it arrives rather than a settled turn. `decisions=false`
        drops `decision`.

        The client sends `tool_result` to answer a `tool_call`, and `say`, `respond`,
        `interrupt`, `instructions` or `close` to act on the session. A `tool_call` is the
        only frame that must be answered: everything else is a report.
      parameters:
        - $ref: "#/components/parameters/SessionID"
        - name: interim
          in: query
          required: false
          description: Also send `hearing` frames, which are the words as they are revised.
          schema:
            type: boolean
            default: false
        - name: decisions
          in: query
          required: false
          description: Send `decision` frames.
          schema:
            type: boolean
            default: true
      responses:
        "101":
          description: The socket is open
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/{modality}/stream:
    get:
      operationId: streamModality
      summary: Route one modality over a socket, for a pipeline running elsewhere
      description: >
        A WebSocket, which OpenAPI cannot describe past the upgrade. This is the routing
        the agent does, offered a piece at a time: a caller running its own pipeline sends
        audio or text and gets transcripts, audio or completions back, and the request is
        failed over and billed exactly as it would be inside a session.

        Every socket opens with a `start` frame naming the target and the cost tags.
        Speech-to-text then takes binary 16 kHz mono PCM and returns `transcript` frames.
        Text-to-speech takes `speak` frames and returns binary audio with
        `synthesis_complete` between utterances: each audio frame opens with a
        little-endian header of a uint32 sample rate, a uint16 channel count and two
        reserved bytes, followed by PCM16 samples. Completions take `respond` frames and
        return `delta` and `complete`. All three report failures as `error` frames and end
        with `closed`.

        Memory and phone are recorded rather than routed, so they are not served here.
      parameters:
        - $ref: "#/components/parameters/Modality"
      responses:
        "101":
          description: The socket is open
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/stats/rollup:
    post:
      operationId: runRollup
      summary: Aggregate request rows into a rollup table
      description: >
        Covers every modality and customer in the window. Idempotent: re-running it over
        the same window recomputes those buckets, so a missed run is fixed by running it
        again.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RollupRequest"
      responses:
        "200":
          description: The rollup completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RollupResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
components:
  securitySchemes:
    CustomerId:
      type: apiKey
      in: header
      name: X-Customer-Id
      description: Trusted customer identifier. Every statistic is keyed by it.
  parameters:
    Modality:
      name: modality
      in: path
      required: true
      description: Which kind of model to route.
      schema:
        $ref: "#/components/schemas/Modality"
    SessionID:
      name: id
      in: path
      required: true
      description: The session, as returned when it was created.
      schema:
        type: string
...
  responses:
    BadRequest:
      description: The request was malformed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: The customer header is missing
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: No such modality, provider or shortcut
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Granularity:
      type: string
      enum: [hourly, daily]
      default: hourly
...
    Provider:
      type: object
      required: [provider, model, languages, realtime, tier, health]
      properties:
        provider:
          type: string
          example: elevenlabs
        model:
          type: string
          example: eleven_flash_v2_5
        languages:
          type: array
          items:
            type: string
        realtime:
          type: boolean
        tier:
          $ref: "#/components/schemas/Tier"
        health:
          $ref: "#/components/schemas/ProviderHealth"
...
    Candidate:
      type: object
      required: [provider, model, health]
      properties:
        provider:
          type: string
        model:
          type: string
        health:
          $ref: "#/components/schemas/ProviderHealth"
    StatsBucket:
      type: object
      required:
        [
          provider,
          model,
          bucket,
          audio_ms_total,
          characters_total,
          input_tokens_total,
          cached_input_tokens_total,
          output_tokens_total,
          cost_micros_total,
          request_count,
          error_count,
        ]
      properties:
        provider:
          type: string
        model:
          type: string
        bucket:
          type: string
          format: date-time
        audio_ms_total:
          type: integer
          format: int64
          description: Billable audio, transcribed or produced.
        characters_total:
          type: integer
          format: int64
          description: Billable text. Zero for providers that bill by audio.
        input_tokens_total:
          type: integer
          format: int64
          description: Prompt tokens read, cached ones included. Zero outside llm.
        cached_input_tokens_total:
          type: integer
          format: int64
          description: The part of the prompt served from the provider's cache.
        output_tokens_total:
          type: integer
          format: int64
          description: Generated tokens, reasoning included. Zero outside llm.
        cost_micros_total:
          type: integer
          format: int64
          description: Millionths of a dollar, priced from the configured rates.
        request_count:
          type: integer
          format: int64
        error_count:
          type: integer
          format: int64
        latency_p50_ms:
          type: number
          format: double
          nullable: true
        latency_p95_ms:
          type: number
          format: double
          nullable: true
        uptime:
          type: number
          format: double
          nullable: true
          description: Successes over total requests in the bucket.
...
    RollupResult:
      type: object
      required: [granularity, buckets_written]
      properties:
        granularity:
          $ref: "#/components/schemas/Granularity"
        buckets_written:
          type: integer
          format: int64